backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
pwdlib[argon2]
PyJWT
pytest
pytest-cov
pytest-asyncio
httpx